entry.c: ensure full index
mark_colliding_entries() scans an index looking for entries that collide
with a given cache entry. It appears to only know about files, and it is
unclear (to me at this moment) what should happen with a sparse
directory in this loop. Use ensure_full_index() until this can be
revisited and tested.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>