fsmonitor: ensure full index
There are several uses of cache_nr in fsmonitor.c, which is a good
indicator that ensure_full_index() is needed until we make the code
aware of sparse directories.

Several of these references are just BUG() checks to see that the
fsmonitor bitmap doesn't have more bits than the index has entries.

fill_fsmonitor_bitmap() iterates through the index entry-by-entry
looking for entries with the CE_REMOVE or without the CE_FSMONITOR_VALID
bits. This seems like we will want to have some awareness of a sparse
directory that the FSMonitor indicates got populated. Save that
awareness for later.

tweak_fsmonitor() has a loop that marks all entries with
CE_FSMONITOR_VALID. I'm not sure that sparse directory entries should
have this bit set, or what would happen if we allowed that. Delay
needing that until we can carefully test the situation.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>