merge-recursive: ensure full index
Even though most of the machinery in merge-recursive.c relies on merging
trees instead of indexes, there are some references to using the index.
In particular, it uses the index as a data structure for unmerged
content. This will require careful integration with sparse indexes to be
aware of sparse directories. Perhaps we will restrict the merge strategy
to use the (currently under review) merge-ort strategy that doesn't use
an index, but we are not ready for that decision. For now, add
ensure_full_index() calls before each index scan, including:

1. get_unmerged() scans index entries that exist in stages other than 0.

2. dir_in_way() is checking if there is a directory/file conflict. It
   must be updated to know that 'path' could be a sparse directory or
   within a sparse directory that needs to be expanded.

3. would_lose_untracked() also looks for a position within the index and
   starts scanning all entries with the exact same name. However, this
   may require expanding a sparse directory, so require a full index for
   now.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>