name-hash.c: ensure full index
name-hash.c contains helper methods for hashing filenames that appear in
the index, including directories that are implied by the filenames. This
seems like a clear area where we want to be aware of sparse directories,
but also we would like to see what we can use as-is without needing to
expand to a full index.

Calls to ensure_full_index() are created in the public methods that
appear to require some additional attention. When we are ready to start
making commands like 'git status' or 'git add' fast for sparse indexes,
we can see if these methods are causing a full index expansion and then
prioritize updating them.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>