rerere: ensure full index
Several methods in rerere.c scan the index for conflicted cache entries.
Most of these should be safe with sparse directory entries, since a
directory conflict must be expanded until finding the base file
conflicts. However, there is one use of index_name_pos() and a scan for
an existing entry in handle_cache() that is similar to some that have
been guarded by ensure_full_index() recently. Guard this one, too, until
it can be tested with a sparse index.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>