resolve-undo: ensure full index
While it is not implemented yet, a sparse index might interact poorly
with the resolve-undo cache. When a merge conflict exists outside of the
sparse-checkout definition, the conflicted files are created in the
working directory despite the sparse-checkout definition. Upon
resolution, we can expect that the sparse-index will attempt to make a
parent directory into a sparse directory entry, removing those paths
from the cache entry list. This makes undoing the merge resolution much
harder, so ensure a full index before doing any of these operations.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>