split-index: ensure full index
The sparse index is not intended to be compatible with a split index.
Trying to write a sparse index is already guarded to not modify a split
index, but we have not guarded the reverse. Add ensure_full_index() to
init_split_index() so any attempt to write a split index from a sparse
index is based on a full index.

Of course, this is likely still wrong, especially if the split index
assumes that the given index matches the on-disk representation of the
base index.

Callers of init_split_index() assume to get a non-NULL result, so the
only way to completely guard against sparse indexes is to die() in the
case of a sparse index. Perhaps that is not necessary, so use
ensure_full_index() instead until proven otherwise.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>