submodule.c: ensure full index
Several methods in submodule.c look for the existence of submodules or
iterate through the list of submodules. These searches are currently
incompatible with a sparse index, so use ensure_full_index() before
searching or scanning the index.

In one case, the repository is known to possibly not be the_repository.
The others are not necessarily linked to a repository, so those still
point to the_repository. Perhaps this area is justification enough to
make an index_state refer to its parent repository so
ensure_full_index() can parse objects in the correct repository without
needing to change too many other callers?

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>