tree.c: ensure full index
The read_tree() method supposedly only looks for stage 1 items, but that
could change in the future. Or maybe not, since the comment warning
about the change was created by af3785d (Optimize "diff --cached"
performance., 2007-08-09).

For now, use ensure_full_index() to be sure we are operating on a full
index until we can test this scenario.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>