grep: ensure full index
The 'git grep' command can scan the current index for certain patterns.
This does not restrict to the populated set when sparse-checkout is
enabled, but maybe it should. However, the loop that scans entries
should be aware of sparse directory entries, so until we can test the
scenario in an interesting way, use ensure_full_index().

Care will also need to be taken to be careful around pathspecs that
could match paths within sparse directories.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>