checkout-index: ensure full index
The checkout_file() method in builtin/checkout-index.c uses
cache_name_pos() to get the current path position from the current
index. This might have strange behavior when the path is inside a sparse
directory, so use ensure_full_index().

While we are here, drop an instance of the antiquated cache_name_pos()
and replace it with index_name_pos(), even though we need to use
the_repository->index as a parameter.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>