rm: ensure full index
The rm builtin removes matching entries from the index. In the presence
of sparse directories, this behavior will change. This behavior change
might be desireable, but should be delayed until we can properly test
and document the behavior. Further, we will likely want to make it
possible to still do the previous behavior, so that will need to be
created as well.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>