read-cache: make refresh_index() sparse aware
The refresh_index() method is called by builtins such as 'git status'
and scans the index for necessary updates. In the case of sparse
directories, we can skip these and move on.

Adjustments must be made for preload_index() as well, since that uses a
threaded scan of the index entries in a similar way.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>