sparse-index: be careful around non-sparse subentries
If a directory is not completely sparse, then we need to be careful.

With a clean status, this doesn't happen when using cone mode.

However, with a non-clean status, we could have a directory that is
sparse in its initial range, but not in some subentry due to either a
non-sparse file that is on-disk or an unmerged conflict.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>