test-lib: test_region looks for trace2 regions
Most test cases can verify Git's behavior using input/output
expectations or changes to the .git directory. However, sometimes we
want to check that Git did or did not run a certain section of code.
This is particularly important for performance-only features that we
want to ensure have been enabled in certain cases.

Add a new 'test_region' function that checks if a trace2 region was
entered during the execution of a certain Git command.

This is similar to test_subcommand in that it greps the trace2 output,
but it doesn't rely on the caller creating a log and instead runs the
command with the proper GIT_TRACE2_* environment variables. It deletes
the trace if it returns with success. To assist with debugging, the
trace file is left on-disk in the case of a failure.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>