read-cache: move format version to context
Before extracting the logic for writing the cache entry table from
do_write_index(), first move the local variable that stores the version
into the write_index_context struct.

The logic for choosing the version previously modified istate->version,
but that seems incorrect. We want to specify the version that we are
about to write, and there is no need to erroneously change the version
for the in-memory index. Thus, we change the logic around loading the
default version or choosing between v2 and v3 to rely on
ctx.format_version instead of istate->version. The only place that
seemed like this could be dangerous was in the init_split_index() call,
but that only allocates data structures and currently ignores
istate->version.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>