t1092: add failing test with cherry-pick conflicts
While integrating the sparse-index with our functional tests in the
Scalar project, I found some failures I was not expecting. Most of the
test failures involved a failure to write a sparse-index when conflicts
exist in the working tree.

If we take this added test and change it to test_expect_success, then
the verbose output fails with this log:

+ diff -u full-checkout-err sparse-index-err
--- full-checkout-err   2021-04-28 19:35:20.664845542 +0000
+++ sparse-index-err    2021-04-28 19:35:20.680845526 +0000
@@ -1,4 +1,6 @@
-error: could not apply df093a0... conflict to pick
-hint: after resolving the conflicts, mark the corrected paths
-hint: with 'git add <paths>' or 'git rm <paths>'
-hint: and commit the result with 'git commit'
+conflict: unmerged (142195f36aeb03b2967436a2cf1f40b6596ad9bb)
+conflict: unmerged (5da849b5c6f00b5292b1a823b29ceb303e98585c)
+warning: unable to update cache-tree, staying full
+warning: failed to convert to a sparse-index
+error: cherry-pick: Unable to write new index file
+fatal: cherry-pick failed

The next change will address what is happening here with regards to the
sparse-index and conflict states.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>