sparse-index: skip indexes with unmerged entries
The cache-tree extension computes tree OIDs for directories within the
index based on the blobs and subtrees within that directory. These OIDs
are used in convert_to_sparse() to construct sparse directory entries
for a sparse index. This is a convenient way to re-use the tree
construction logic within cache-tree.c:update_one().

This is a particularly critical path right now because the sparse-index
is frequently converted to a full one in memory, then converted back
before writing to disk. Operations like 'git checkout' use this
translation instead of swapping the sparse directory entries directly
with a known tree OID.

However, if the index has unmerged entries due to a merge conflict, the
update_cache_tree() method fails to update the cache-tree data
structure.

In general, unmerged entries should not be a blocker for the
sparse-index, since we will never collapse them into a sparse-directory
entry. However, if we do not have a cache-tree data structure, then the
algorithm in convert_to_sparse() fails to work.

This change is a fix in that it avoids runtime failures when merge
conflicts occur, but it is not a good long-term fix because it simply
fails to convert to a sparse index when merge conflicts exist.

A more appropriate long-term fix might be to allow the construction of a
cache-tree data structure that has invalid tree entries for those that
cover unmerged entries. This might be a difficult effort, so this quick
fix exists to solve the existing problem until that more careful
solution can be built and tested carefully.

Add tests that demonstrate this behavior across three builtins:
cherry-pick, merge, and rebase.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>