fsmonitor--daemon: trace response token always
TODO squash this into the original code.

The trivial response did not include tracing of the response token in
the same way that a full response (including a concrete list of changes)
did. Copy these lines from earlier in the method to also exist below the
send_trivial_response label.

Signed-off-by: Jeff Hostetler <jeffhost@microsoft.com>
Signed-off-by: Derrick Stolee <dstolee@microsoft.com>