commit: integrate with sparse-index
Update 'git commit' to allow using the sparse-index in memory without
expanding to a full one. The only place that had an ensure_full_index()
call was in cache_tree_update(). The recursive algorithm for
update_one() was already updated in 2de37c536 (cache-tree: integrate
with sparse directory entries, 2021-03-03) to handle sparse directory
entries in the index.

Most of this change involves testing different command-line options that
allow specifying which on-disk changes should be included in the commit.
This includes no options (only take currently-staged changes), -a (take
all tracked changes), and --include (take a list of specific changes).
To simplify testing that these options do not expand the index, update
the test that previously verified that 'git status' does not expand the
index with a helper method, ensure_not_expanded().

This allows 'git commit' to operate much faster when the sparse-checkout
cone is much smaller than the full list of files at HEAD.

Here are the relevant lines from p2000-sparse-operations.sh:

Test                                            HEAD~1            HEAD
----------------------------------------------------------------------------------------
2000.14: git commit -a -m A (full-index-v3)     1.36(1.08+0.16)   1.36(1.08+0.17) +0.0%
2000.15: git commit -a -m A (full-index-v4)     1.24(1.05+0.13)   1.25(1.08+0.12) +0.8%
2000.16: git commit -a -m A (sparse-index-v3)   2.18(2.07+0.19)   0.04(0.04+0.06) -98.2%
2000.17: git commit -a -m A (sparse-index-v4)   2.19(2.10+0.16)   0.04(0.06+0.04) -98.2%

It is important to compare the full-index case to the sparse-index case,
so the improvement for index version v4 is actually a 96% improvement in
this synthetic example.

In a real repository with over two million files at HEAD and 60,000
files in the sparse-checkout definition, the time for 'git commit -a'
went from 2.61 seconds to 134ms. I compared this to the result if the
index only contained the paths in the sparse-checkout definition and
found the theoretical optimum to be 120ms, so the out-of-cone paths only
add a 12% overhead.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>