sparse-index: silently return when not using cone-mode patterns
This commit is necessary becaue of the previous change that turned on
the sparse-index by default. Some cases in t1091 now enable the sparse
index and hit different corner cases around cone mode patterns.

Also silently return if there are no cache entries, which is a simple
case: there are no paths to make sparse!

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>