sparse-index: silently return when cache tree fails
If cache_tree_update() returns a non-zero value, then it could not
create the cache tree likely due to a merge conflict. If we remove our
dependence on the cache tree within convert_to_sparse(), then we could
still recover from this scenario and have a sparse index. Since we are
already returning early, let's return silently to avoid making it seem
like we failed to write the index at all.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>