sparse-index: fix crash in status
Copy the `index_state->dir_hash` back to the real istate
after expanding a sparse index.

A crash was observed in `git status` during some hashmap lookups with
corrupted hashmap entries.  During an index expansion, new cache-entries
are added to the `index_state->name_hash` and the `dir_hash` in a
temporary `index_state` variable `full`.  However, only the `name_hash`
hashmap from this temp variable was copied back into the real `istate`
variable.  The original copy of the `dir_hash` was incorrectly preserved.
If the table in the `full->dir_hash` hashmap were realloced, the
stale version (in `istate`) would be corrupted.

Signed-off-by: Jeff Hostetler <jeffhost@microsoft.com>