t1092: test ignored files outside of cone
When a sparse directory leaves the sparse-checkout definition, we leave
the ignored files in place. This then requires commands like 'git
status' to expand a sparse index to find contained '.gitignore' files
within the sparse directory.

This test is marked as an expected failure, and the failures happen
within the ensure_not_expanded lines.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>