t7519: rewrite sparse index test
This test was very fragile because it shared an index across sparse and
non-sparse behavior. Since that expansion and contraction could cause
the index to lose its FS Monitor bitmap and token, behavior is fragile
to changes in 'git sparse-checkout set'.

Rewrite the test to use two clones of the original repo: full and
sparse. This allows us to also keep the test files (actual, expect,
trace2.txt) out of the repos we are testing with 'git status'.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>